[Kinetics] Make ReactionFactory internal
