+++ src/kinetics/Reaction.cpp
@@ -6,7 +6,7 @@
 // at https://cantera.org/license.txt for license and copyright information.
 
 #include "cantera/kinetics/Reaction.h"
-#include "cantera/kinetics/ReactionFactory.h"
+#include "ReactionFactory.h"
 #include "cantera/kinetics/ReactionRateFactory.h"
 #include "cantera/kinetics/FalloffFactory.h"
 #include "cantera/kinetics/Kinetics.h"
@@ -1259,6 +1259,105 @@ CustomFunc1Reaction::CustomFunc1Reaction(const AnyMap& node, const Kinetics& kin
     } else {
         setRate(newReactionRate(type()));
     }
+}
+
+bool isThreeBody(const Reaction& R)
+{
+    // detect explicitly specified collision partner
+    size_t found = 0;
+    for (const auto& reac : R.reactants) {
+        auto prod = R.products.find(reac.first);
+        if (prod != R.products.end() &&
+            trunc(reac.second) == reac.second && trunc(prod->second) == prod->second) {
+            // candidate species with integer stoichiometric coefficients on both sides
+            found += 1;
+        }
+    }
+    if (found != 1) {
+        return false;
+    }
+
+    // ensure that all reactants have integer stoichiometric coefficients
+    size_t nreac = 0;
+    for (const auto& reac : R.reactants) {
+       if (trunc(reac.second) != reac.second) {
+           return false;
+       }
+       nreac += static_cast<size_t>(reac.second);
+    }
+
+    // ensure that all products have integer stoichiometric coefficients
+    size_t nprod = 0;
+    for (const auto& prod : R.products) {
+       if (trunc(prod.second) != prod.second) {
+           return false;
+       }
+       nprod += static_cast<size_t>(prod.second);
+    }
+
+    // either reactant or product side involves exactly three species
+    return (nreac == 3) || (nprod == 3);
+}
+
+unique_ptr<Reaction> newReaction(const std::string& type)
+{
+    AnyMap rxn_node;
+    Kinetics kin;
+    unique_ptr<Reaction> R(ReactionFactory::factory()->create(type, rxn_node, kin));
+    return R;
+}
+
+unique_ptr<Reaction> newReaction(const XML_Node& rxn_node)
+{
+    std::string type = toLowerCopy(rxn_node["type"]);
+
+    // Modify the reaction type for interface reactions which contain
+    // electrochemical reaction data
+    if (rxn_node.child("rateCoeff").hasChild("electrochem")
+        && (type == "edge" || type == "surface")) {
+        type = "electrochemical";
+    }
+
+    if (!(ReactionFactoryXML::factory()->exists(type))) {
+        throw CanteraError("newReaction",
+            "Unknown reaction type '" + rxn_node["type"] + "'");
+    }
+    if (type.empty()) {
+        // Reaction type is not specified
+        // See if this is a three-body reaction with a specified collision partner
+        ElementaryReaction2 testReaction;
+        setupReaction(testReaction, rxn_node);
+        if (isThreeBody(testReaction)) {
+            type = "three-body";
+        }
+    }
+    Reaction* R = ReactionFactoryXML::factory()->create(type, rxn_node);
+    return unique_ptr<Reaction>(R);
+}
+
+unique_ptr<Reaction> newReaction(const AnyMap& rxn_node, const Kinetics& kin)
+{
+    std::string type = "elementary";
+    size_t nDim = kin.thermo(kin.reactionPhaseIndex()).nDim();
+    if (rxn_node.hasKey("type")) {
+        type = rxn_node["type"].asString();
+    } else if (nDim == 3) {
+        // Reaction type is not specified
+        // See if this is a three-body reaction with a specified collision partner
+        ElementaryReaction2 testReaction;
+        parseReactionEquation(testReaction, rxn_node["equation"].asString(),
+                              rxn_node, &kin);
+        if (isThreeBody(testReaction)) {
+            type = "three-body";
+        }
+    }
+
+    if (!(ReactionFactory::factory()->exists(type))) {
+        throw InputFileError("ReactionFactory::newReaction", rxn_node["type"],
+            "Unknown reaction type '{}'", type);
+    }
+    Reaction* R = ReactionFactory::factory()->create(type, rxn_node, kin);
+    return unique_ptr<Reaction>(R);
 }
 
 Arrhenius2 readArrhenius(const XML_Node& arrhenius_node)
