+++ src/zeroD/ConstPressureReactor.cpp
@@ -8,6 +8,7 @@
 #include "cantera/zeroD/Wall.h"
 #include "cantera/kinetics/Kinetics.h"
 #include "cantera/thermo/SurfPhase.h"
+#include "cantera/base/utilities.h"
 
 using namespace std;
