+++ src/zeroD/IdealGasConstPressureReactor.cpp
@@ -8,6 +8,7 @@
 #include "cantera/zeroD/ReactorNet.h"
 #include "cantera/kinetics/Kinetics.h"
 #include "cantera/thermo/ThermoPhase.h"
+#include "cantera/base/utilities.h"
 
 using namespace std;
