+++ src/zeroD/IdealGasReactor.cpp
@@ -8,6 +8,7 @@
 #include "cantera/zeroD/Wall.h"
 #include "cantera/kinetics/Kinetics.h"
 #include "cantera/thermo/ThermoPhase.h"
+#include "cantera/base/utilities.h"
 
 using namespace std;
