+++ test/kinetics/kineticsFromYaml.cpp
@@ -5,7 +5,6 @@
 #include "cantera/kinetics/GasKinetics.h"
 #include "cantera/thermo/SurfPhase.h"
 #include "cantera/kinetics/KineticsFactory.h"
-#include "cantera/kinetics/ReactionFactory.h"
 #include "cantera/thermo/ThermoFactory.h"
 #include "cantera/base/Array.h"
